refactor(workspace): refactoring code to be more declarative
